(gh-64) Cherry-pick acceptance changes from 9c8670926
This commit tears out the execution of the 'teardown/common/099_Archive_Logs.rb'
as part of the Beaker post_suite, since that step was only relevant
to jenkins.

It adds a presuite step used by the GHA action to run configure_type_defaults_on()
to get the .ssh/environment setup for beaker. Normally beaker-puppet
would do this, but it's not capable of installing openvox-agent, so the
standup_cluster plan is handling agent installation instead. The
configure_type_defaults_on() step is important because Beaker's use of
ssh as root relies on the .ssh/environment file supplying PATH to the
puppet binaries.