(gh-29) Switch acceptance workflow to dispatch
Now that kvm_automation_tooling/action.yaml accepts
parameters for openvox_version and can test pre-release builds from the
artifacts.overlookinfratech.com server, switch to a workflow_dispatch.

Until such a time as there is a build step that would build and push
packages to an artifacts server, a pr or branch push event isn't
meaningful to test against aside from validating the gha acceptance.yaml
workflow itself, since we would just be validating an already released
package unreleated to the code in the pr or, for most cases outside of a
release tag, the branch.