(gh-64) Fix two acceptance tests for openvox7
The validate_vendored_openssl and validate_vendored_ruby tests were
failing when run against openvox7. The openssl test was expecting
openssl 3 instead of the v1 shipped with openvox7. And the ruby test was
attempting a gem update --system which was pulling the latest
rubygem-update which is incompatible with the ruby 2.7 shipped with
openvox7.

Presumably the suite was never run for the puppet 7 collection, or I'm
missing something else -- or these tests were specifically skipped?

Either way, it was simple enough to add a version test for these two,
allowing them to pass by testing for openssl v1, and tightening up the
gem update --system to 3.4.22 which was the last rubygems available for
ruby 2.7.

(cherry picked from commit 5b94b5f3cdf76e97880172b2a488524b41f43d08)