(maint) Use ruby 3 for gha checks.yaml
Updating the checks.yaml run of rubocop from ruby 2.7 to 3.3.

The 7.x branch depends on openvoxproject/vanagon@main which currently
requires a minimum of Ruby 3.2 due to git gem dependencies.

While 7.x bundles 2.7 in the package, there's no reason to hamstring the
local tests to run 2.7 since it's testing build configuration which is
going to be run with Ruby 3 anyway.